requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
numpy>=1.21.0
pillow>=8.3.0

# Data processing
datasets>=2.0.0
opencv-python>=4.5.0
albumentations>=1.2.0

# Training and evaluation
tensorboard>=2.9.0
wandb>=0.12.0
tqdm>=4.64.0
scikit-learn>=1.1.0

# Configuration and utilities
omegaconf>=2.2.0
hydra-core>=1.2.0
einops>=0.4.0
timm>=0.6.0

# Optional: For distributed training
deepspeed>=0.7.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0